prop types refactoring
