+++ packages/scandipwa/src/component/ProductGalleryBaseImage/ProductGalleryBaseImage.container.js
@@ -13,7 +13,7 @@ import PropTypes from 'prop-types';
 import { Component } from 'react';
 import { withRouter } from 'react-router';
 
-import { LocationType } from 'Type/Common';
+import { LocationType } from 'Type/Router';
 import media, { PRODUCT_MEDIA } from 'Util/Media/Media';
 
 import ProductGallery from './ProductGalleryBaseImage.component';
