+++ packages/scandipwa/src/component/PureForm/Field/Field.component.js
@@ -18,6 +18,7 @@ import FieldFile from 'Component/PureForm/FieldFile';
 import { FieldNumberContainer } from 'Component/PureForm/FieldNumber/FieldNumber.container';
 import FieldSelectContainer from 'Component/PureForm/FieldSelect/FieldSelect.container';
 import { MixType } from 'Type/Common';
+import { OptionType } from 'Type/Field';
 
 import './Field.style';
 
@@ -31,11 +32,11 @@ export class Field extends PureComponent {
         // Field attributes
         type: PropTypes.oneOf(Object.values(FIELD_TYPE)).isRequired,
         attr: PropTypes.object.isRequired,
-        events: PropTypes.object.isRequired,
+        events: PropTypes.objectOf(PropTypes.func).isRequired,
         isDisabled: PropTypes.bool.isRequired,
         setRef: PropTypes.func.isRequired,
         mix: MixType.isRequired,
-        options: PropTypes.array.isRequired,
+        options: PropTypes.arrayOf(OptionType).isRequired,
 
         // Validation
         showErrorAsLabel: PropTypes.bool.isRequired,
