+++ packages/scandipwa/src/route/CmsPage/CmsPage.container.js
@@ -19,7 +19,7 @@ import { updateMeta } from 'Store/Meta/Meta.action';
 import { changeNavigationState } from 'Store/Navigation/Navigation.action';
 import { TOP_NAVIGATION_TYPE } from 'Store/Navigation/Navigation.reducer';
 import { setBigOfflineNotice } from 'Store/Offline/Offline.action';
-import { LocationType, MatchType } from 'Type/Common';
+import { LocationType, MatchType } from 'Type/Router';
 import { scrollToTop } from 'Util/Browser';
 import history from 'Util/History';
 import { debounce } from 'Util/Request';
