+++ packages/scandipwa/src/route/MenuPage/MenuPage.container.js
@@ -19,7 +19,7 @@ import Menu from 'Component/Menu';
 import { updateMeta } from 'Store/Meta/Meta.action';
 import { changeNavigationState } from 'Store/Navigation/Navigation.action';
 import { TOP_NAVIGATION_TYPE } from 'Store/Navigation/Navigation.reducer';
-import { HistoryType } from 'Type/Common';
+import { HistoryType } from 'Type/Router';
 import { appendWithStoreCode } from 'Util/Url';
 
 /** @namespace Route/MenuPage/Container/mapStateToProps */
