+++ packages/scandipwa/src/type/Common.js
@@ -20,27 +20,6 @@ export const MixType = PropTypes.shape({
     ]))
 });
 
-export const LocationType = PropTypes.shape({
-    pathname: PropTypes.string,
-    search: PropTypes.string,
-    state: PropTypes.object,
-    hash: PropTypes.string,
-    key: PropTypes.string
-});
-
-export const HistoryType = {
-    length: PropTypes.number,
-    action: PropTypes.string,
-    location: LocationType
-};
-
-export const MatchType = PropTypes.shape({
-    path: PropTypes.string,
-    url: PropTypes.string,
-    params: PropTypes.object,
-    isExact: PropTypes.bool
-});
-
 export const ChildrenType = PropTypes.oneOfType([
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node
